bot/requirement.txt
fastapi>=0.63.0
uvicorn>=0.13.4
python-twitter>=3.5
pysocks>=1.7.1